+ Adding first performance test.
Someone [linked to a stackoverflow][so] post that mentioned that another library
that "has much better performance for read operations" than ours. This got me
looking at the performance of ActsAsTree tonight. I'm adding the first (in what
I hope is a series) performance test to the library.

Starting with the obviously-common case of deeply nested relationships, I
assumed the read performance was linear. Within and error threshold of 0.95 the
reads _are_ linear. I don't have equivalent tests for the other library, but
at least we have a jumping off point to ensure ActsAsTree performance can be
tested, validated and improved.

[so]: http://stackoverflow.com/questions/1319809/recursive-model-in-rails#comment1153380_1319859